PIM-6353: Renamed stuff
