+++ src/Pim/Bundle/CatalogBundle/spec/Elasticsearch/Indexer/ProductIndexerSpec.php
@@ -11,8 +11,8 @@ use Akeneo\Component\StorageUtils\Remover\RemoverInterface;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Elasticsearch\Indexer\ProductIndexer;
 use Pim\Component\Catalog\Model\ProductInterface;
+use Pim\Component\Catalog\Normalizer\Indexing\Product\ProductNormalizer;
 use Pim\Component\Catalog\Normalizer\Indexing\ProductAndModel\ProductModelNormalizer;
-use Pim\Component\Catalog\Normalizer\Indexing\ProductNormalizer;
 use Prophecy\Argument;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
