+++ src/Pim/Component/Catalog/spec/Normalizer/Indexing/Product/FamilyNormalizerSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Component\Catalog\Normalizer\Indexing\Product;
+namespace spec\Pim\Component\Catalog\Normalizer\Indexing\ProductValue;
 
 use PhpSpec\ObjectBehavior;
 use Pim\Component\Catalog\Model\FamilyInterface;
