+++ src/Pim/Component/Catalog/spec/Normalizer/Indexing/Product/ProductNormalizerSpec.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace spec\Pim\Component\Catalog\Normalizer\Indexing;
+namespace spec\Pim\Component\Catalog\Normalizer\Indexing\Product;
 
 use PhpSpec\ObjectBehavior;
 use Pim\Component\Catalog\Model\ProductInterface;
-use Pim\Component\Catalog\Normalizer\Indexing\ProductNormalizer;
+use Pim\Component\Catalog\Normalizer\Indexing\Product\ProductNormalizer;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 class ProductNormalizerSpec extends ObjectBehavior
